decArgo_20170905_013a commit

+++ decArgo_soft/soft/util/sub/read_file_traj_3_1.m
@@ -175,11 +175,11 @@ else
             end
             if (size(data, 2) == 1)
                paramDataNbDimList = [paramDataNbDimList 1];
-               paramData = [paramData data];
+               paramData = [paramData double(data)];
                paramQcData = [paramQcData dataQc];
             else
                paramDataNbDimList = [paramDataNbDimList size(data, 1)];
-               paramData = [paramData data'];
+               paramData = [paramData double(data)'];
                paramQcData = [paramQcData dataQc];
             end
             
@@ -198,11 +198,11 @@ else
                   dataAdjQc = netcdf.getVar(fCdf, netcdf.inqVarID(fCdf, paramAdjQcName));
                   if (size(dataAdj, 2) == 1)
                      paramAdjDataNbDimList = [paramAdjDataNbDimList 1];
-                     paramAdjData = [paramAdjData dataAdj];
+                     paramAdjData = [paramAdjData double(dataAdj)];
                      paramAdjQcData = [paramAdjQcData dataAdjQc];
                   else
                      paramAdjDataNbDimList = [paramAdjDataNbDimList size(dataAdj, 1)];
-                     paramAdjData = [paramAdjData dataAdj'];
+                     paramAdjData = [paramAdjData double(dataAdj)'];
                      paramAdjQcData = [paramAdjQcData dataAdjQc'];
                   end
                   %                end
